small edits to step model, took our re-written data values.
changed "datetime" inputs for create and update forms into "datetime-local" because it creates an automatic filter for an entry.

fixed user registration validation static method on the models and then added it into the user controller post registration route to be called on.

Added these validation flashes into the html of registration page.

Created logout user button, controller route, logout flash on login page.

added validation to see if user_id in session for logged in user to be logged in in order to see any other page.